GPS functionality moved to seperate viewmodel
